feat : ImageManager 구현
